+++ tests/testthat.R
@@ -18,6 +18,10 @@ fl <- system.file("TripleTOF-SWATH", "PestMix1_SWATH.mzML",
                   package = "msdata")
 sps_dia <- Spectra(fl)
 
+fl <- system.file("TripleTOF-SWATH", "PestMix1_DDA.mzML",
+                  package = "msdata")
+sps_dda <- Spectra(fl)
+
 sciex_hd5 <- backendInitialize(MsBackendHdf5Peaks(),
                                data = spectraData(sciex_mzr),
                                hdf5path = tempdir())