+++ NEWS.md
@@ -1,5 +1,10 @@
 # Spectra 1.11
 
+## Changes in 1.11.3
+
+- Add `filterPrecursorMaxIntensity` function.
+- Add `filterPrecursorIsotopes` function.
+
 ## Changes in 1.11.2
 
 - Add `deisotopeSpectra` and `reduceSpectra` functions.